Break out code editor widgets
