+++ src/Interface/Modules/Python/CMakeLists.txt
@@ -35,12 +35,14 @@ SET(Interface_Modules_Python_FORMS
 SET(Interface_Modules_Python_HEADERS
   PythonObjectForwarderDialog.h
   InterfaceWithPythonDialog.h
+  CodeEditorWidgets.h
   share.h
 )
 
 SET(Interface_Modules_Python_SOURCES
   PythonObjectForwarderDialog.cc
   InterfaceWithPythonDialog.cc
+  CodeEditorWidgets.cc
 )
 
 QT_WRAP_UI(Interface_Modules_Python_FORMS_HEADERS "${Interface_Modules_Python_FORMS}")
