+++ src/Interface/Modules/Python/InterfaceWithPythonDialog.h
@@ -53,105 +53,6 @@ private:
   class CodeEditor* pythonCodePlainTextEdit_;
 };
 
-class CodeEditor : public QPlainTextEdit
-{
-  Q_OBJECT
-
-public:
-  explicit CodeEditor(QWidget *parent = nullptr);
-
-  void lineNumberAreaPaintEvent(QPaintEvent *event);
-  int lineNumberAreaWidth();
-
-public Q_SLOTS:
-  void insertSpecialCodeBlock();
-protected:
-  void resizeEvent(QResizeEvent *event) override;
-
-private Q_SLOTS:
-  void updateLineNumberAreaWidth(int newBlockCount);
-  void highlightCurrentLine();
-  void updateLineNumberArea(const QRect &, int);
-  void matchParentheses();
-
-private:
-  QWidget* lineNumberArea_;
-  class Highlighter* highlighter_;
-  void createParenthesisSelection(int pos);
-  bool matchLeftParenthesis(QTextBlock currentBlock, int index, int numRightParentheses);
-  bool matchRightParenthesis(QTextBlock currentBlock, int index, int numLeftParentheses);
-};
-
-class LineNumberArea : public QWidget
-{
-public:
-  explicit LineNumberArea(CodeEditor *editor) : QWidget(editor) 
-  {
-    codeEditor = editor;
-  }
-
-  QSize sizeHint() const override 
-  {
-    return QSize(codeEditor->lineNumberAreaWidth(), 0);
-  }
-
-protected:
-  void paintEvent(QPaintEvent *event) override 
-  {
-    codeEditor->lineNumberAreaPaintEvent(event);
-  }
-
-private:
-  CodeEditor *codeEditor;
-};
-
-class Highlighter : public QSyntaxHighlighter
-{
-  Q_OBJECT
-
-public:
-  Highlighter(QTextDocument* parent = nullptr);
-
-protected:
-  void highlightBlock(const QString &text) override;
-
-private:
-  struct HighlightingRule
-  {
-    QRegExp pattern;
-    QTextCharFormat format;
-  };
-  QVector<HighlightingRule> highlightingRules;
-
-  QRegExp commentStartExpression;
-  QRegExp commentEndExpression;
-
-  QTextCharFormat keywordFormat;
-  QTextCharFormat classFormat;
-  QTextCharFormat singleLineCommentFormat;
-  QTextCharFormat multiLineCommentFormat;
-  QTextCharFormat quotationFormat;
-  QTextCharFormat functionFormat;
-
-  void highlightBlockParens(const QString& text);
-};
-
-struct ParenthesisInfo
-{
-  char character;
-  int position;
-};
-
-class TextBlockData : public QTextBlockUserData
-{
-public:
-  TextBlockData();
-  QVector<ParenthesisInfo *> parentheses();
-  void insert(ParenthesisInfo *info);
-private:
-  QVector<ParenthesisInfo *> m_parentheses;
-};
-
 }
 }
